+++ lib/src/widgets/nes_check_box.dart
@@ -1,3 +1,5 @@
+// ignore_for_file: avoid_positional_boolean_parameters
+
 import 'package:flutter/material.dart';
 import 'package:nes_ui/nes_ui.dart';